refactor: 🔨 Update folder structure
