+++ sigma/validators/sigmahq/data/sigmahq_data.py
