+++ tests/sigmahq/author/test_sigmahq_author_existence_validator.py
