+++ tests/sigmahq/condition/test_sigmahq_missing_asterisk_condition_validator.py
