+++ tests/sigmahq/condition/test_sigmahq_ofselection_condition_validator.py
