+++ tests/sigmahq/condition/test_sigmahq_ofthem_condition_validator.py
