+++ tests/sigmahq/correlation/test_sigmahq_correlation_rules_minimum.py
