+++ tests/sigmahq/description/test_sigmahq_link_in_description_validator.py
