+++ tests/sigmahq/detection/test_sigmahq_category_eventid_validator.py
