+++ tests/sigmahq/detection/test_sigmahq_category_windows_provider_name_validator.py
