+++ tests/sigmahq/detection/test_sigmahq_unsupported_regex_group_construct_validator.py
