+++ tests/sigmahq/falsepositive/test_sigmahq_falsepositives_banned_word_validator.py
