+++ tests/sigmahq/falsepositive/test_sigmahq_falsepositives_capital_validator.py
