+++ tests/sigmahq/falsepositive/test_sigmahq_falsepositives_typo_word_validator.py
