+++ tests/sigmahq/field/test_sigmahq_invalid_all_modifier_validator.py
