+++ tests/sigmahq/field/test_sigmahq_invalid_fieldname_validator.py
