+++ tests/sigmahq/field/test_sigmahq_redundant_field_validator,.py
@@ -1,6 +1,5 @@
 from sigma.rule import SigmaRule
 from sigma.correlations import SigmaCorrelationRule
-from sigma.types import SigmaRegularExpression
 
 from sigma.validators.sigmahq.field import (
     SigmahqRedundantFieldIssue,
