+++ tests/sigmahq/references/test_sigmahq_mitre_link_validator.py
