+++ tests/sigmahq/tags/test_sigmahq_tags_detection_validator.py
