+++ tests/sigmahq/tags/test_sigmahq_tags_techniques_without_tactics_validator.py
