+++ tests/sigmahq/title/test_sigmahq_title_case_validator.py
